Change int parameter to size_t

+++ std-udp-recv/include/FrameStats.hpp
@@ -9,7 +9,7 @@
 class FrameStats {
     const std::string detector_name_;
     const int module_id_;
-    const int n_packets_per_frame_;
+    const size_t n_packets_per_frame_;
     const size_t stats_time_;
 
     int frames_counter_;
@@ -23,7 +23,7 @@ class FrameStats {
 public:
     FrameStats(std::string detector_name,
                const int module_id,
-               const int n_packets_per_frame,
+               const size_t n_packets_per_frame,
                const size_t stats_time);
     void record_stats(uint64_t n_missing_packets);
 };
